refcto the conversions out
