+++ ferritin-core/src/core/atomcollection.rs
@@ -411,128 +411,6 @@ fn match_bond(bond_int: i32) -> BondOrder {
     }
 }
 
-impl From<&PSEData> for AtomCollection {
-    fn from(pse_data: &PSEData) -> Self {
-        let mols = pse_data.get_molecule_data();
-
-        // Pymol: most of the descriptive data is there
-        let atoms: Vec<&ferritin_pymol::pymolparsing::parsing::AtomInfo> =
-            mols.iter().flat_map(|mol| mol.atom.iter()).collect();
-
-        // Pymol: coord sets are maintained seperately.
-        let coord_sets: Vec<&ferritin_pymol::pymolparsing::parsing::CoordSet> =
-            mols.iter().flat_map(|mol| mol.coord_set.iter()).collect();
-
-        let coords: Vec<[f32; 3]> = coord_sets
-            .iter()
-            .flat_map(|c| c.get_coords_as_vec())
-            .collect();
-
-        // Pymol: most of the descriptive data is there
-        let pymol_bonds: Vec<&ferritin_pymol::pymolparsing::parsing::Bond> =
-            mols.iter().flat_map(|mol| mol.bond.iter()).collect();
-
-        let bonds = pymol_bonds
-            .iter()
-            .map(|b| Bond {
-                atom1: b.index_1,
-                atom2: b.index_2,
-                order: match_bond(b.order),
-            })
-            .collect();
-
-        // pull out specific fields
-        let (res_names, res_ids, chain_ids, is_hetero, elements, atom_names): (
-            Vec<String>,
-            Vec<i32>,
-            Vec<String>,
-            Vec<bool>,
-            Vec<Element>,
-            Vec<String>,
-        ) = atoms
-            .iter()
-            .map(|a| {
-                (
-                    a.resn.to_string(),
-                    a.resv,
-                    a.chain.to_string(),
-                    a.is_hetatm,
-                    a.elem,
-                    a.name.to_string(),
-                )
-            })
-            .multiunzip();
-
-        AtomCollection {
-            size: atoms.len(),
-            coords,
-            res_names,
-            res_ids,
-            chain_ids,
-            is_hetero,
-            elements,
-            atom_names,
-            bonds: Some(bonds),
-        }
-    }
-}
-
-impl From<&PDB> for AtomCollection {
-    // the PDB API requires us to iterate:
-    // PDB --> Chain --> Residue --> Atom if we want data from all.
-    // Here we collect all the data in one go and return an AtomCollection
-    fn from(pdb_data: &PDB) -> Self {
-        let (coords, is_hetero, atom_names, res_ids, res_names, elements, chain_ids): (
-            Vec<[f32; 3]>,
-            Vec<bool>,
-            Vec<String>,
-            Vec<i32>,
-            Vec<String>,
-            Vec<Element>,
-            Vec<String>,
-        ) = izip!(pdb_data.chains().flat_map(|chain| {
-            let chain_id = chain.id().to_string();
-            chain.residues().flat_map(move |residue| {
-                let (res_number, _insertion_code) = residue.id();
-                let res_id = res_number as i32;
-                let res_name = residue.name().unwrap_or_default().to_string();
-                let chain_id = chain_id.clone();
-                residue.atoms().filter_map(move |atom| {
-                    atom.element().map(|element| {
-                        let (x, y, z) = atom.pos();
-                        (
-                            [x as f32, y as f32, z as f32],
-                            atom.hetero(),
-                            atom.name().to_string(),
-                            res_id,
-                            res_name.clone(),
-                            element,
-                            chain_id.clone(),
-                        )
-                    })
-                })
-            })
-        }))
-        .multiunzip();
-
-        let mut ac = AtomCollection {
-            size: coords.len(),
-            coords,
-            res_ids,
-            res_names,
-            is_hetero,
-            elements,
-            chain_ids,
-            atom_names,
-            bonds: None,
-        };
-
-        // adds bonds
-        ac.connect_via_residue_names();
-        ac
-    }
-}
-
 /// Bond
 #[derive(Debug, PartialEq)]
 pub struct Bond {
